Make model dropdown always visible.
