+++ src/js/settings/components/provider-settings/googleai-gemini.js
@@ -2,12 +2,15 @@
  * WordPress dependencies
  */
 import { useSelect, useDispatch } from '@wordpress/data';
+import { SelectControl } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import { STORE_NAME } from '../../data/store';
+import { SettingsRow } from '../settings-row';
 import { GoogleAISettings } from './googleai';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Component for Google AI (Gemini) Provider settings.
@@ -26,15 +29,67 @@ export const GoogleAIGeminiSettings = ( { isConfigured = false } ) => {
 	const { setProviderSettings } = useDispatch( STORE_NAME );
 	const onChange = ( data ) => setProviderSettings( providerName, data );
 
+	const models = [];
+
+	// Convert providerSettings.models to an array from an object.
+	if (
+		providerSettings?.models &&
+		! Array.isArray( providerSettings.models )
+	) {
+		for ( const [ key, value ] of Object.entries(
+			providerSettings.models
+		) ) {
+			models.push( { label: value, value: key } );
+		}
+	} else {
+		models.push( {
+			label: __( '-- Choose Model --', 'classifai' ),
+			value: '',
+		} );
+	}
+
+	const ModelDescription = () => (
+		<>
+			{ __(
+				'Choose the model you want to use for requests.',
+				'classifai'
+			) }{ ' ' }
+			{ __(
+				'Not sure which model to use? You can find more details on models',
+				'classifai'
+			) }{ ' ' }
+			<a
+				title={ __( 'Learn more about models', 'classifai' ) }
+				href="https://ai.google.dev/gemini-api/docs/models"
+				target="_blank"
+				rel="noopener noreferrer"
+			>
+				{ __( 'here', 'classifai' ) }
+			</a>
+			.
+		</>
+	);
+
 	return (
 		<>
 			{ ! isConfigured && (
 				<GoogleAISettings
 					providerSettings={ providerSettings }
 					onChange={ onChange }
-					providerName={ providerName }
 				/>
 			) }
+			<SettingsRow
+				label={ __( 'Model', 'classifai' ) }
+				description={ <ModelDescription /> }
+			>
+				<SelectControl
+					id={ `${ providerName }_model` }
+					onChange={ ( value ) => onChange( { model: value } ) }
+					value={ providerSettings?.model || '' }
+					options={ models }
+					__nextHasNoMarginBottom
+				/>
+			</SettingsRow>
 		</>
 	);
 };
