+++ src/js/settings/components/provider-settings/googleai.js
@@ -1,7 +1,5 @@
-import {
-	__experimentalInputControl as InputControl, // eslint-disable-line @wordpress/no-unsafe-wp-apis
-	SelectControl,
-} from '@wordpress/components';
+// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
+import { __experimentalInputControl as InputControl } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import { SettingsRow } from '../settings-row';
 
@@ -13,15 +11,10 @@ import { SettingsRow } from '../settings-row';
  * @param {Object}   props                  Component props.
  * @param {Object}   props.providerSettings The provider settings.
  * @param {Function} props.onChange         Function to call when the provider settings change.
- * @param {string}   props.providerName     The provider name. Default is 'googleai'.
  *
  * @return {React.ReactElement} GoogleAISettings component.
  */
-export const GoogleAISettings = ( {
-	providerSettings,
-	onChange,
-	providerName = 'googleai',
-} ) => {
+export const GoogleAISettings = ( { providerSettings, onChange } ) => {
 	const Description = () => (
 		<>
 			{ __( "Don't have an Google AI (Gemini API) key?", 'classifai' ) }{ ' ' }
@@ -34,47 +27,6 @@ export const GoogleAISettings = ( {
 				{ __( 'Get an API key', 'classifai' ) }
 			</a>{ ' ' }
 			{ __( 'now.', 'classifai' ) }
-		</>
-	);
-
-	const models = [];
-
-	// Convert providerSettings.models to an array from an object.
-	if (
-		providerSettings?.models &&
-		! Array.isArray( providerSettings.models )
-	) {
-		for ( const [ key, value ] of Object.entries(
-			providerSettings.models
-		) ) {
-			models.push( { label: value, value: key } );
-		}
-	} else {
-		models.push( {
-			label: __( '-- Choose Model --', 'classifai' ),
-			value: '',
-		} );
-	}
-
-	const ModelDescription = () => (
-		<>
-			{ __(
-				'Choose the model you want to use for requests.',
-				'classifai'
-			) }{ ' ' }
-			{ __(
-				'Not sure which model to use? You can find more details on models',
-				'classifai'
-			) }{ ' ' }
-			<a
-				title={ __( 'Learn more about models', 'classifai' ) }
-				href="https://ai.google.dev/gemini-api/docs/models"
-				target="_blank"
-				rel="noopener noreferrer"
-			>
-				{ __( 'here', 'classifai' ) }
-			</a>
-			.
 		</>
 	);
 
@@ -91,20 +43,6 @@ export const GoogleAISettings = ( {
 					onChange={ ( value ) => onChange( { api_key: value } ) }
 				/>
 			</SettingsRow>
-			{ 'googleai_gemini_api' === providerName && (
-				<SettingsRow
-					label={ __( 'Model', 'classifai' ) }
-					description={ <ModelDescription /> }
-				>
-					<SelectControl
-						id={ `${ providerName }_model` }
-						onChange={ ( value ) => onChange( { model: value } ) }
-						value={ providerSettings?.model || '' }
-						options={ models }
-						__nextHasNoMarginBottom
-					/>
-				</SettingsRow>
-			) }
 		</>
 	);
 };
